Break out each submit page into components
